refactor(#40): modify chats dto file naming
