+++ src/chat/dto/lookup-chat-dto.ts
@@ -1,7 +1,7 @@
 import { PickType } from '@nestjs/swagger';
-import { ChatsDto } from './chats.dto';
+import { ChatDto } from './chat.dto';
 
-export class LookupChatsDto extends PickType(ChatsDto, [
+export class LookupChatsDto extends PickType(ChatDto, [
   'content',
   'createdAt',
   'seenUsers',
